requirements.txt
pdfplumber
python-bidi
backoff
python-dotenv==1.0.1
beautifulsoup4==4.12.3
boto3==1.34.133
openai==1.40.1
pymongo==4.7.3
Requests==2.32.2
tiktoken==0.6.0
pandas==2.2.2
openpyxl==3.1.2
regex==2024.4.28
rapidfuzz==3.9.3
backoff==2.2.1
jinja2==3.1.5
nats-py==2.9.0
httpx==0.27.2
Python-IO
mongomock